fixed circular dependency with GeometryCollection
